requirements.txt
addict==2.4.0
aiofiles==24.1.0
gensim==4.3.0
google_cloud_aiplatform==1.49.0
gradio==4.43.0
huggingface_hub==0.22.2
ipdb==0.13.13
matplotlib==3.8.0
numpy==1.24.4
onnxruntime==1.19.2
opencv_python==4.9.0.80
Pillow==9.3.0
pycocotools==2.0.7
PyYAML==6.0.1
PyYAML==6.0
PyYAML==6.0.2
Requests==2.32.3
scipy==1.11.4
setuptools==68.2.2
skimage==0.0
supervision==0.23.0
termcolor==2.4.0
timm==1.0.9
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.35.2
vertexai==1.49.0
yapf==0.40.2
